+++ .idea/libraries/Maven__io_zipkin_brave_brave_instrumentation_http_5_10_1.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: io.zipkin.brave:brave-instrumentation-http:5.10.1">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/zipkin/brave/brave-instrumentation-http/5.10.1/brave-instrumentation-http-5.10.1.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/io/zipkin/brave/brave-instrumentation-http/5.10.1/brave-instrumentation-http-5.10.1-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/zipkin/brave/brave-instrumentation-http/5.10.1/brave-instrumentation-http-5.10.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
