+++ .idea/libraries/Maven__io_zipkin_reporter2_zipkin_sender_amqp_client_2_12_1.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: io.zipkin.reporter2:zipkin-sender-amqp-client:2.12.1">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/zipkin/reporter2/zipkin-sender-amqp-client/2.12.1/zipkin-sender-amqp-client-2.12.1.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/io/zipkin/reporter2/zipkin-sender-amqp-client/2.12.1/zipkin-sender-amqp-client-2.12.1-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/zipkin/reporter2/zipkin-sender-amqp-client/2.12.1/zipkin-sender-amqp-client-2.12.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
