+++ .idea/libraries/Maven__io_zipkin_zipkin2_zipkin_2_19_3.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: io.zipkin.zipkin2:zipkin:2.19.3">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/zipkin/zipkin2/zipkin/2.19.3/zipkin-2.19.3.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/io/zipkin/zipkin2/zipkin/2.19.3/zipkin-2.19.3-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/zipkin/zipkin2/zipkin/2.19.3/zipkin-2.19.3-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
