+++ .idea/libraries/Maven__org_springframework_cloud_spring_cloud_starter_sleuth_2_2_2_RELEASE.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.springframework.cloud:spring-cloud-starter-sleuth:2.2.2.RELEASE">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/cloud/spring-cloud-starter-sleuth/2.2.2.RELEASE/spring-cloud-starter-sleuth-2.2.2.RELEASE.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/cloud/spring-cloud-starter-sleuth/2.2.2.RELEASE/spring-cloud-starter-sleuth-2.2.2.RELEASE-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/springframework/cloud/spring-cloud-starter-sleuth/2.2.2.RELEASE/spring-cloud-starter-sleuth-2.2.2.RELEASE-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
