+++ clientui/src/main/java/com/livefox/clientui/configuration/SleuthConfig.java
@@ -1,0 +1,11 @@
+package com.livefox.clientui.configuration;
+
+import brave.sampler.Sampler;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SleuthConfig {
+    public Sampler defaultSampler(){
+        return Sampler.ALWAYS_SAMPLE;
+    }
+}
