+++ clientui/src/main/java/com/livefox/clientui/controller/ClientController.java
@@ -2,6 +2,8 @@ package com.livefox.clientui.controller;
 
 import com.livefox.clientui.bean.VideoBean;
 import com.livefox.clientui.proxies.VideoProxy;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -15,10 +17,13 @@ public class ClientController {
     @Autowired
     VideoProxy videoProxy;
 
+    Logger log = LoggerFactory.getLogger(this.getClass());
+
     @RequestMapping("/")
     public String home(Model model){
         List<VideoBean> videos= videoProxy.listVideo();
         model.addAttribute("videos",videos);
+        log.info("Request home page");
         return"Home";
     }
 }
