+++ eureka-server/src/main/java/com/livefox/eurekaserver/configuration/SleuthConfig.java
@@ -1,0 +1,11 @@
+package com.livefox.eurekaserver.configuration;
+
+import brave.sampler.Sampler;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SleuthConfig {
+    public Sampler defaultSampler(){
+        return Sampler.ALWAYS_SAMPLE;
+    }
+}
