+++ microservice_video/microservice-video/src/main/java/com/livefox/video/controller/VideoController.java
@@ -1,9 +1,11 @@
 package com.livefox.video.controller;
 
-import com.livefox.video.configurations.ApplicationPropertiesConfigurations;
+import com.livefox.video.configuration.ApplicationPropertiesConfigurations;
 import com.livefox.video.exception.VideoNotFoundException;
 import com.livefox.video.model.Video;
 import com.livefox.video.repository.VideoRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,12 +21,16 @@ public class VideoController {
     @Autowired
     ApplicationPropertiesConfigurations appProperties;
 
+    Logger log = LoggerFactory.getLogger(this.getClass());
+
     @GetMapping(value = "/video")
     public List<Video> listVideo(){
         List<Video> video = videoRepository.findAll();
         if(video.isEmpty()) throw new VideoNotFoundException("Aucune video disponible");
 
         List<Video> listLimite = video.subList(0,appProperties.getLimitDeVideos());
+
+        log.info("Find all videos");
         return listLimite ;
     }
 
@@ -40,6 +46,7 @@ public class VideoController {
         videoRepository.save(video3);
         videoRepository.save(video4);
         videoRepository.save(video5);
+        log.info("Autofilled MYSQL Databased");
 
     }
 
@@ -48,21 +55,25 @@ public class VideoController {
         Video video = videoRepository.findById(id);
 
         if(video == null) throw new VideoNotFoundException("Cette video n'existe pas");
+        log.info("Request video by id: " + video.getId());
         return video;
     }
 
     @PostMapping(value = "/video/add")
     public void addVideo (@Valid @RequestBody Video video){
         videoRepository.save(video);
+        log.info("add a video: " + video.toString());
     }
 
     @GetMapping(value = "/video/delete/{id]")
     public void deleteVideo(@PathVariable int id){
+        log.info("delete the video id: " + id);
         videoRepository.deleteById(id);
     }
 
     @PostMapping(value = "/video/update/{id}")
     public void updateVideo (@RequestBody Video video){
+        log.info("update a video: " + video.toString());
         videoRepository.save(video);
     }
 }
